refactor: split jobs module and move probe/split pure helpers
Break the jobs monofile into store, queue, and executor submodules with
stable re-exports. Move probe_or_rebuild and pool item split into sizing
so load/pool budgeting stays free of ONNX session types.
