--- crates/gigastt/src/server/jobs.rs
@@ -1,1755 +1,0 @@
-//! Asynchronous job queue for long-file and batch transcription.
-//!
-//! The queue is intentionally decoupled from the HTTP surface via the
-//! [`JobStore`] trait so a persistent backend (e.g. SQLite) can plug in
-//! later without touching the handlers.
-
-use arc_swap::ArcSwap;
-use axum::body::Bytes;
-use serde::{Deserialize, Serialize};
-use std::collections::{HashMap, VecDeque};
-use std::future::Future;
-use std::pin::Pin;
-use std::str::FromStr;
-use std::sync::Arc;
-use std::sync::atomic::{AtomicBool, AtomicU64, Ordering};
-
-use parking_lot::Mutex;
-
-use super::config::RuntimeLimits;
-use super::http::ExportParams;
-
-/// Object-safe boxed future returned by [`JobStore`] methods.
-pub type JobStoreFuture<'a, T> = Pin<Box<dyn Future<Output = T> + Send + 'a>>;
-
-/// Lifecycle status of a transcription job.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum JobStatus {
-    /// Waiting for a worker slot.
-    Queued,
-    /// Currently holding a triplet and transcribing.
-    Processing,
-    /// Finished successfully; result is available.
-    Done,
-    /// Failed after exhausting retries.
-    Failed,
-    /// Cancelled by the client or by shutdown.
-    Cancelled,
-}
-
-impl JobStatus {
-    /// Whether the job has reached a terminal state.
-    pub fn is_terminal(self) -> bool {
-        matches!(
-            self,
-            JobStatus::Done | JobStatus::Failed | JobStatus::Cancelled
-        )
-    }
-}
-
-/// Server-sent event emitted by `GET /v1/jobs/{id}/events`.
-#[derive(Debug, Clone, Serialize)]
-#[serde(tag = "type", rename_all = "snake_case")]
-pub enum JobEvent {
-    /// Progress estimate while processing.
-    Progress {
-        /// Approximate fraction complete, 0–100.
-        percent: u32,
-        /// Seconds of audio considered processed so far.
-        processed_seconds: f64,
-    },
-    /// Job completed successfully.
-    Done,
-    /// Job failed.
-    Failed {
-        /// Sanitized error message (no paths or model internals).
-        error: String,
-    },
-    /// Job was cancelled.
-    Cancelled,
-}
-
-impl JobEvent {
-    /// Whether this event ends the stream.
-    pub fn is_terminal(&self) -> bool {
-        matches!(
-            self,
-            JobEvent::Done | JobEvent::Failed { .. } | JobEvent::Cancelled
-        )
-    }
-}
-
-/// Public status response for `GET /v1/jobs/{id}`.
-#[derive(Debug, Clone, Serialize)]
-pub struct JobStatusResponse {
-    pub job_id: String,
-    pub status: JobStatus,
-    pub processed_seconds: f64,
-    pub percent: u32,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub error: Option<String>,
-}
-
-/// Build a public status view from a stored job.
-pub(crate) fn job_status_response(job: &Job) -> JobStatusResponse {
-    let percent = if job.total_seconds > 0.0 {
-        ((job.processed_seconds / job.total_seconds) * 100.0) as u32
-    } else {
-        0
-    };
-    JobStatusResponse {
-        job_id: job.id.clone(),
-        status: job.status,
-        processed_seconds: job.processed_seconds,
-        percent,
-        error: job.error.clone(),
-    }
-}
-
-/// A transcription job.
-#[derive(Debug, Clone)]
-pub struct Job {
-    pub id: String,
-    pub status: JobStatus,
-    /// Raw uploaded audio bytes.
-    pub body: Bytes,
-    /// Export / post-processing parameters captured at submission.
-    pub params: ExportParams,
-    pub created_at: f64,
-    pub updated_at: f64,
-    pub processed_seconds: f64,
-    /// Total audio duration in seconds, set once the body is decoded.
-    pub total_seconds: f64,
-    /// Number of execution attempts made so far.
-    pub attempts: u32,
-    /// Populated when status becomes `Done`.
-    pub result: Option<gigastt_core::inference::TranscribeResult>,
-    /// Why speakers were or were not labeled, recorded when the run finishes and
-    /// only when `?diarization=true` was submitted. `GET /v1/jobs/{id}/result`
-    /// turns it into the same capability notice the synchronous endpoint
-    /// attaches, so an async job cannot answer with silently empty speaker
-    /// fields either. `None` when diarization was not requested.
-    pub diarization: Option<gigastt_core::inference::DiarizationOutcome>,
-    /// Populated when status becomes `Failed`.
-    pub error: Option<String>,
-    /// Active SSE listeners.
-    pub event_channels: Vec<tokio::sync::mpsc::UnboundedSender<JobEvent>>,
-    /// Cooperative-cancellation flag for the in-flight run. The executor sets it
-    /// when it begins inference; `DELETE /v1/jobs/{id}` flips it so the engine
-    /// releases its pooled triplet within one window instead of transcribing the
-    /// whole file. `None` while queued and after a terminal state. Purely
-    /// in-memory (never serialized): it is a live handle, not job metadata.
-    pub abort: Option<Arc<AtomicBool>>,
-}
-
-/// Upper bound on simultaneous SSE listeners per job. Dead channels are
-/// pruned on subscribe and on broadcast, but broadcasts can be ≥500 ms apart
-/// (and rarer for a queued job), so a connect/disconnect flood could otherwise
-/// accumulate `UnboundedSender`s faster than they get cleaned up.
-pub(crate) const MAX_JOB_EVENT_SUBSCRIBERS: usize = 32;
-
-impl Job {
-    /// Create a new queued job.
-    pub fn queued(body: Bytes, params: ExportParams) -> Self {
-        let now = gigastt_core::inference::now_timestamp();
-        Self {
-            id: uuid::Uuid::now_v7().to_string(),
-            status: JobStatus::Queued,
-            body,
-            params,
-            created_at: now,
-            updated_at: now,
-            processed_seconds: 0.0,
-            total_seconds: 0.0,
-            attempts: 0,
-            result: None,
-            diarization: None,
-            error: None,
-            event_channels: Vec::new(),
-            abort: None,
-        }
-    }
-
-    /// Register an SSE listener, pruning closed channels first. When the
-    /// subscriber cap is reached, the oldest listener is evicted (its stream
-    /// ends) so a connect/disconnect flood cannot grow the list without bound.
-    pub(crate) fn subscribe(&mut self, tx: tokio::sync::mpsc::UnboundedSender<JobEvent>) {
-        self.event_channels.retain(|tx| !tx.is_closed());
-        if self.event_channels.len() >= MAX_JOB_EVENT_SUBSCRIBERS {
-            self.event_channels.remove(0);
-        }
-        self.event_channels.push(tx);
-    }
-}
-
-/// Persistence boundary for jobs. Handlers talk to this trait; the in-memory
-/// implementation is the default, but a SQLite-backed store can be dropped in.
-pub trait JobStore: Send + Sync + 'static {
-    /// Persist a new job and return its id.
-    fn create<'a>(&'a self, job: Job) -> JobStoreFuture<'a, anyhow::Result<String>>;
-    /// Return a clone of the job, if it exists.
-    fn get<'a>(&'a self, id: &str) -> JobStoreFuture<'a, anyhow::Result<Option<Job>>>;
-    /// Apply an in-place mutation.
-    fn update<'a>(
-        &'a self,
-        id: &str,
-        f: Box<dyn FnOnce(&mut Job) + Send>,
-    ) -> JobStoreFuture<'a, anyhow::Result<()>>;
-    /// Pop the oldest queued job id whose status is still `Queued`.
-    fn next_queued<'a>(&'a self) -> JobStoreFuture<'a, anyhow::Result<Option<String>>>;
-    /// Push a job id to the back of the queue (used after a retryable failure).
-    fn requeue<'a>(&'a self, id: &str) -> JobStoreFuture<'a, anyhow::Result<()>>;
-    /// Whether the store has reached its capacity limit.
-    fn is_full<'a>(&'a self) -> JobStoreFuture<'a, bool>;
-}
-
-/// In-memory FIFO job store with TTL eviction.
-pub struct InMemoryJobStore {
-    limits: RuntimeLimits,
-    jobs: Mutex<HashMap<String, Job>>,
-    queue: Mutex<VecDeque<String>>,
-}
-
-impl InMemoryJobStore {
-    /// Create a new store with the given limits.
-    pub fn new(limits: RuntimeLimits) -> Self {
-        Self {
-            limits,
-            jobs: Mutex::new(HashMap::new()),
-            queue: Mutex::new(VecDeque::new()),
-        }
-    }
-
-    /// Evict terminal jobs whose TTL has expired. Must be called with both locks held.
-    fn evict_expired_locked(&self, jobs: &mut HashMap<String, Job>, queue: &mut VecDeque<String>) {
-        let ttl = self.limits.jobs_ttl_secs;
-        if ttl == 0 {
-            return;
-        }
-        let now = gigastt_core::inference::now_timestamp();
-        let expired: Vec<String> = jobs
-            .iter()
-            .filter(|(_, j)| j.status.is_terminal() && now - j.updated_at > ttl as f64)
-            .map(|(id, _)| id.clone())
-            .collect();
-        for id in expired {
-            jobs.remove(&id);
-            queue.retain(|x| x != &id);
-        }
-    }
-
-    /// Total bytes of buffered uploads currently resident across all jobs.
-    /// Terminal jobs release their body (`Bytes::new()`), so this is the live
-    /// upload footprint — the quantity `jobs_max_bytes` bounds. O(jobs), and
-    /// `jobs_max` already caps the count, so this is a short walk.
-    fn resident_body_bytes(jobs: &HashMap<String, Job>) -> usize {
-        jobs.values().map(|j| j.body.len()).sum()
-    }
-
-    /// Whether the store is at capacity by count OR by resident upload bytes.
-    /// Folding the byte budget in here means an over-budget queue produces the
-    /// exact same 429 + `Retry-After` backpressure as a count-full one, with no
-    /// new error type. Bounds live upload RAM to `jobs_max_bytes` plus at most
-    /// one `body_limit_bytes` (the job that crosses the line is admitted, the
-    /// next is refused), mirroring how the count cap admits exactly `jobs_max`.
-    fn at_capacity(&self, jobs: &HashMap<String, Job>) -> bool {
-        jobs.len() >= self.limits.jobs_max
-            || Self::resident_body_bytes(jobs) >= self.limits.jobs_max_bytes
-    }
-}
-
-impl JobStore for InMemoryJobStore {
-    fn create<'a>(&'a self, job: Job) -> JobStoreFuture<'a, anyhow::Result<String>> {
-        Box::pin(async move {
-            let mut jobs = self.jobs.lock();
-            let mut queue = self.queue.lock();
-            self.evict_expired_locked(&mut jobs, &mut queue);
-            // Race backstop behind the `is_full` gate in `submit_job` (which
-            // returns 429 + Retry-After): rejects on either the count or the
-            // byte budget so two concurrent submits can't both slip past the gate.
-            if self.at_capacity(&jobs) {
-                return Err(anyhow::anyhow!("job store is full"));
-            }
-            let id = job.id.clone();
-            jobs.insert(id.clone(), job);
-            queue.push_back(id.clone());
-            Ok(id)
-        })
-    }
-
-    fn get<'a>(&'a self, id: &str) -> JobStoreFuture<'a, anyhow::Result<Option<Job>>> {
-        let id = id.to_owned();
-        Box::pin(async move {
-            let jobs = self.jobs.lock();
-            Ok(jobs.get(&id).cloned())
-        })
-    }
-
-    fn update<'a>(
-        &'a self,
-        id: &str,
-        f: Box<dyn FnOnce(&mut Job) + Send>,
-    ) -> JobStoreFuture<'a, anyhow::Result<()>> {
-        let id = id.to_owned();
-        Box::pin(async move {
-            let mut jobs = self.jobs.lock();
-            let Some(job) = jobs.get_mut(&id) else {
-                return Err(anyhow::anyhow!("job not found"));
-            };
-            f(job);
-            job.updated_at = gigastt_core::inference::now_timestamp();
-            Ok(())
-        })
-    }
-
-    fn next_queued<'a>(&'a self) -> JobStoreFuture<'a, anyhow::Result<Option<String>>> {
-        Box::pin(async move {
-            let mut jobs = self.jobs.lock();
-            let mut queue = self.queue.lock();
-            self.evict_expired_locked(&mut jobs, &mut queue);
-            while let Some(id) = queue.pop_front() {
-                if let Some(job) = jobs.get(&id)
-                    && matches!(job.status, JobStatus::Queued)
-                {
-                    return Ok(Some(id));
-                }
-            }
-            Ok(None)
-        })
-    }
-
-    fn requeue<'a>(&'a self, id: &str) -> JobStoreFuture<'a, anyhow::Result<()>> {
-        let id = id.to_owned();
-        Box::pin(async move {
-            let mut queue = self.queue.lock();
-            queue.push_back(id);
-            Ok(())
-        })
-    }
-
-    fn is_full<'a>(&'a self) -> JobStoreFuture<'a, bool> {
-        Box::pin(async move {
-            let mut jobs = self.jobs.lock();
-            let mut queue = self.queue.lock();
-            self.evict_expired_locked(&mut jobs, &mut queue);
-            self.at_capacity(&jobs)
-        })
-    }
-}
-
-/// Executor abstraction so unit tests can run the queue without loading ONNX.
-pub trait JobExecution: Send + Sync {
-    /// Run one transcription attempt. The executor may update progress via the
-    /// store and should return an error for retryable failures.
-    fn execute(
-        &self,
-        id: &str,
-        store: Arc<dyn JobStore>,
-        body: Bytes,
-        params: ExportParams,
-    ) -> impl std::future::Future<Output = anyhow::Result<gigastt_core::inference::TranscribeResult>>
-    + Send;
-}
-
-/// In-memory FIFO job queue. Spawns `concurrency` workers that pull from the
-/// store, run the executor, and retry up to `max_retries` on transient failures.
-pub struct JobQueue {
-    store: Arc<dyn JobStore>,
-    semaphore: Arc<tokio::sync::Semaphore>,
-    max_retries: u32,
-    shutdown: tokio_util::sync::CancellationToken,
-}
-
-impl JobQueue {
-    /// Create a new queue. `concurrency` is clamped to at least 1.
-    pub fn new(
-        store: Arc<dyn JobStore>,
-        concurrency: usize,
-        max_retries: u32,
-        shutdown: tokio_util::sync::CancellationToken,
-    ) -> Arc<Self> {
-        Arc::new(Self {
-            store,
-            semaphore: Arc::new(tokio::sync::Semaphore::new(concurrency.max(1))),
-            max_retries,
-            shutdown,
-        })
-    }
-
-    /// Spawn worker tasks onto the server drain [`TaskTracker`] so graceful
-    /// shutdown waits for in-flight jobs (same lane as WS / SSE / REST).
-    /// Each worker owns a clone of `executor`. Call once after constructing
-    /// the queue.
-    pub fn spawn<E>(&self, executor: E, tracker: &tokio_util::task::TaskTracker)
-    where
-        E: JobExecution + Clone + Send + Sync + 'static,
-    {
-        let permits = self.semaphore.available_permits();
-        for _ in 0..permits {
-            let worker = JobWorker {
-                store: self.store.clone(),
-                semaphore: self.semaphore.clone(),
-                max_retries: self.max_retries,
-                shutdown: self.shutdown.clone(),
-                executor: executor.clone(),
-            };
-            tracker.spawn(worker.run());
-        }
-    }
-
-    /// Mark all queued jobs as cancelled. Called during graceful shutdown.
-    pub async fn cancel_all_queued(&self) {
-        loop {
-            let Ok(Some(id)) = self.store.next_queued().await else {
-                break;
-            };
-            let _ = self
-                .store
-                .update(
-                    &id,
-                    Box::new(|j| {
-                        j.status = JobStatus::Cancelled;
-                        j.updated_at = gigastt_core::inference::now_timestamp();
-                    }),
-                )
-                .await;
-            broadcast_event(&*self.store, &id, JobEvent::Cancelled).await;
-        }
-    }
-}
-
-struct JobWorker<E> {
-    store: Arc<dyn JobStore>,
-    semaphore: Arc<tokio::sync::Semaphore>,
-    max_retries: u32,
-    shutdown: tokio_util::sync::CancellationToken,
-    executor: E,
-}
-
-impl<E: JobExecution + 'static> JobWorker<E> {
-    async fn run(self) {
-        loop {
-            if self.shutdown.is_cancelled() {
-                break;
-            }
-            let permit = match tokio::time::timeout(
-                std::time::Duration::from_secs(1),
-                self.semaphore.clone().acquire_owned(),
-            )
-            .await
-            {
-                Ok(Ok(p)) => p,
-                _ => continue,
-            };
-            let Some(id) = self.store.next_queued().await.unwrap_or(None) else {
-                drop(permit);
-                tokio::time::sleep(std::time::Duration::from_millis(100)).await;
-                continue;
-            };
-
-            let _ = self
-                .store
-                .update(
-                    &id,
-                    Box::new(|j| {
-                        j.status = JobStatus::Processing;
-                        j.attempts += 1;
-                    }),
-                )
-                .await;
-            broadcast_event(
-                &*self.store,
-                &id,
-                JobEvent::Progress {
-                    percent: 0,
-                    processed_seconds: 0.0,
-                },
-            )
-            .await;
-
-            let store = self.store.clone();
-            let body = match self.store.get(&id).await {
-                Ok(Some(job)) => job.body,
-                _ => {
-                    drop(permit);
-                    continue;
-                }
-            };
-            let params = match self.store.get(&id).await {
-                Ok(Some(job)) => job.params,
-                _ => {
-                    drop(permit);
-                    continue;
-                }
-            };
-
-            let result = self
-                .executor
-                .execute(&id, store.clone(), body, params)
-                .await;
-            drop(permit);
-
-            // If the job was cancelled while running, discard the result.
-            let cancelled = self
-                .store
-                .get(&id)
-                .await
-                .ok()
-                .flatten()
-                .map(|j| matches!(j.status, JobStatus::Cancelled))
-                .unwrap_or(false);
-            if cancelled {
-                continue;
-            }
-
-            match result {
-                Ok(res) => {
-                    let total = res.duration_s;
-                    let _ = self
-                        .store
-                        .update(
-                            &id,
-                            Box::new(move |j| {
-                                j.status = JobStatus::Done;
-                                j.result = Some(res);
-                                j.processed_seconds = total;
-                                // Release the upload: a terminal job never needs
-                                // its body again, and holding it for the full TTL
-                                // would keep up to jobs_max × body-limit of dead
-                                // audio in RAM.
-                                j.body = Bytes::new();
-                            }),
-                        )
-                        .await;
-                    broadcast_event(&*self.store, &id, JobEvent::Done).await;
-                }
-                Err(e) => {
-                    let attempts = self
-                        .store
-                        .get(&id)
-                        .await
-                        .ok()
-                        .flatten()
-                        .map(|j| j.attempts)
-                        .unwrap_or(0);
-                    let retryable = is_retryable_error(&e) && attempts <= self.max_retries;
-                    if retryable {
-                        // Keep the body: the requeued job needs it for the
-                        // next attempt. It is released when the job reaches a
-                        // terminal state.
-                        let _ = self
-                            .store
-                            .update(
-                                &id,
-                                Box::new(|j| {
-                                    j.status = JobStatus::Queued;
-                                }),
-                            )
-                            .await;
-                        let _ = self.store.requeue(&id).await;
-                        broadcast_event(
-                            &*self.store,
-                            &id,
-                            JobEvent::Progress {
-                                percent: 0,
-                                processed_seconds: 0.0,
-                            },
-                        )
-                        .await;
-                    } else {
-                        let sanitized = sanitize_job_error(&e);
-                        let _ = self
-                            .store
-                            .update(
-                                &id,
-                                Box::new({
-                                    let sanitized = sanitized.clone();
-                                    move |j| {
-                                        j.status = JobStatus::Failed;
-                                        j.error = Some(sanitized);
-                                        // Same release as the Done path: a
-                                        // failed job is terminal and no longer
-                                        // needs its upload.
-                                        j.body = Bytes::new();
-                                    }
-                                }),
-                            )
-                            .await;
-                        broadcast_event(&*self.store, &id, JobEvent::Failed { error: sanitized })
-                            .await;
-                    }
-                }
-            }
-        }
-    }
-}
-
-/// Send an event to all active listeners, pruning dead channels. Terminal
-/// events are fire-and-forget: the channel is dropped afterwards so the SSE
-/// stream ends naturally.
-pub(crate) async fn broadcast_event(store: &dyn JobStore, id: &str, event: JobEvent) {
-    let channels = match store.get(id).await {
-        Ok(Some(job)) => job.event_channels,
-        _ => return,
-    };
-    let terminal = event.is_terminal();
-    let mut keep = Vec::new();
-    for tx in channels {
-        if tx.send(event.clone()).is_ok() && !terminal {
-            keep.push(tx);
-        }
-    }
-    let _ = store
-        .update(
-            id,
-            Box::new(move |j| {
-                j.event_channels = keep;
-            }),
-        )
-        .await;
-}
-
-fn is_retryable_error(e: &anyhow::Error) -> bool {
-    // Only a panic in the inference thread is transient: the triplet is
-    // recovered via catch_unwind and the same input may well succeed on a fresh
-    // worker, so it is worth another attempt.
-    //
-    // An `inference_timeout` is deliberately NOT retryable. A file that is too
-    // slow for the limit will be exactly as slow next time, so retrying it just
-    // burns (max_retries + 1) × the timeout on a job that can never pass while
-    // the rest of the queue waits behind it. Fail it once instead.
-    let s = format!("{e:#}");
-    s.contains("panicked")
-}
-
-fn sanitize_job_error(e: &anyhow::Error) -> String {
-    // Preserve the typed length rejection so a client can tell "too long" (raise
-    // `--max-audio-secs` or split) from "corrupt" — the same distinction the REST
-    // path answers with 413. The blocking result reaches us via
-    // `anyhow::Error::from`, so the concrete variant survives the downcast.
-    if let Some(gigastt_core::error::GigasttError::AudioTooLong {
-        observed_secs,
-        limit_secs,
-    }) = e.downcast_ref::<gigastt_core::error::GigasttError>()
-    {
-        return format!(
-            "Audio too long: {observed_secs:.0}s exceeds the maximum of {limit_secs:.0}s."
-        );
-    }
-    let msg = format!("{e:#}");
-    if msg.contains("inference_timeout") {
-        "Inference timed out.".into()
-    } else if msg.contains("Invalid audio") {
-        "Failed to decode audio file. Check format.".into()
-    } else {
-        "Transcription failed.".into()
-    }
-}
-
-/// Production executor: decodes audio to size the job, runs inference against
-/// the batch pool, and emits timer-based progress events.
-#[derive(Clone)]
-pub struct RealJobExecutor {
-    engine: Arc<ArcSwap<gigastt_core::inference::Engine>>,
-    limits: Arc<ArcSwap<RuntimeLimits>>,
-    /// Server shutdown signal. A fired token flips the per-run abort flag so an
-    /// in-flight job releases its triplet at the next window during a drain.
-    shutdown: tokio_util::sync::CancellationToken,
-}
-
-impl RealJobExecutor {
-    /// Create an executor bound to the live engine, runtime limits, and the
-    /// server shutdown token.
-    pub fn new(
-        engine: Arc<ArcSwap<gigastt_core::inference::Engine>>,
-        limits: Arc<ArcSwap<RuntimeLimits>>,
-        shutdown: tokio_util::sync::CancellationToken,
-    ) -> Self {
-        Self {
-            engine,
-            limits,
-            shutdown,
-        }
-    }
-}
-
-impl JobExecution for RealJobExecutor {
-    async fn execute(
-        &self,
-        id: &str,
-        store: Arc<dyn JobStore>,
-        body: Bytes,
-        params: ExportParams,
-    ) -> anyhow::Result<gigastt_core::inference::TranscribeResult> {
-        let engine = self.engine.load_full();
-        let limits = self.limits.load();
-
-        // Size the progress bar from the container header when it declares a
-        // duration (WAV / FLAC / M4A / OGG do). A probe reads O(header) bytes
-        // and decodes nothing, so we skip the full O(T) decode this used to run
-        // purely to throw the samples away — the engine re-decodes `body` for
-        // the actual transcript below. That eliminates one wasted full decode
-        // per job: its wall time and its transient f32 buffer (measured ~0.9 s
-        // and ~73 MB on a 20-min 48 kHz file). Peak RSS is roughly unchanged —
-        // the old pre-decode buffer was already freed before the engine decode,
-        // so the two never coexisted — but the redundant CPU pass is gone.
-        const TARGET_SAMPLE_RATE: f64 = 16_000.0;
-        let probed = tokio::task::spawn_blocking({
-            let body = body.clone();
-            move || gigastt_core::inference::audio::probe_duration_bytes(body)
-        })
-        .await
-        .map_err(|e| anyhow::anyhow!("audio probe task panicked: {e}"))?
-        .ok()
-        .flatten();
-
-        let total_seconds = match probed {
-            Some(seconds) => seconds,
-            None => {
-                // The header declares no usable duration (often a raw MP3
-                // stream): fall back to the full decode so the job keeps its
-                // progress bar. Wrap the decoder in catch_unwind so a malformed
-                // file cannot be retried as a transient inference panic; the
-                // samples are only used for their count and then dropped.
-                let samples = tokio::task::spawn_blocking({
-                    let body = body.clone();
-                    move || {
-                        std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-                            gigastt_core::inference::audio::decode_audio_bytes_shared(body)
-                        }))
-                    }
-                })
-                .await
-                .map_err(|e| anyhow::anyhow!("audio decode task panicked: {e}"))?
-                .map_err(|_| anyhow::anyhow!("Invalid audio: decoder panicked"))?
-                .map_err(|e| anyhow::anyhow!("Invalid audio: {e:#}"))?;
-                samples.len() as f64 / TARGET_SAMPLE_RATE
-            }
-        };
-        let _ = store
-            .update(id, Box::new(move |j| j.total_seconds = total_seconds))
-            .await;
-
-        // Validate per-request variant / knob overrides before holding a triplet.
-        if let Some(requested) = params.variant.as_deref() {
-            let matches = gigastt_core::model::ModelVariant::from_str(requested)
-                .map(|v| v == engine.variant())
-                .unwrap_or(false);
-            if !matches {
-                return Err(anyhow::anyhow!("Requested model variant is not loaded"));
-            }
-        }
-        let overrides = super::http::overrides_from_export_params(&params);
-        if let Err(e) = engine.validate_overrides(&overrides) {
-            return Err(anyhow::anyhow!("Invalid input: {}", e.message()));
-        }
-        let hotwords = super::http::hotwords_from_export_params(&params);
-        if let Some(ref hw) = hotwords
-            && let Err(e) = engine.validate_hotwords(hw)
-        {
-            return Err(anyhow::anyhow!("Invalid input: {}", e.message()));
-        }
-
-        // Cooperative cancellation + real progress. `abort` is flipped by
-        // `DELETE /v1/jobs/{id}` or a shutdown drain; `progress` carries the
-        // engine's cumulative processed 16 kHz sample count out of the blocking
-        // decode. Register `abort` on the job so the cancel handler can reach it.
-        let abort = Arc::new(AtomicBool::new(false));
-        let progress = Arc::new(AtomicU64::new(0));
-        let _ = store
-            .update(id, {
-                let abort = abort.clone();
-                Box::new(move |j| {
-                    // Honour a cancel that raced in between the worker marking
-                    // this job Processing and this registration: seed the flag
-                    // from the current status so the run still stops at its first
-                    // window instead of ignoring the already-recorded cancel.
-                    if matches!(j.status, JobStatus::Cancelled) {
-                        abort.store(true, Ordering::Relaxed);
-                    }
-                    j.abort = Some(abort);
-                })
-            })
-            .await;
-        let shutdown = self.shutdown.clone();
-
-        // Real per-window progress updater: mirrors the engine's processed-sample
-        // count into the store and SSE stream every 500 ms. No RTF guess — the
-        // bar tracks audio actually decoded, monotonically. Same interval and
-        // `JobEvent` shape as before, so the SSE wire contract is unchanged.
-        let progress_cancel = tokio_util::sync::CancellationToken::new();
-        let progress_handle = {
-            let store = store.clone();
-            let id = id.to_string();
-            let cancel = progress_cancel.clone();
-            let total = total_seconds;
-            let progress = progress.clone();
-            tokio::spawn(async move {
-                let mut interval = tokio::time::interval(std::time::Duration::from_millis(500));
-                interval.set_missed_tick_behavior(tokio::time::MissedTickBehavior::Delay);
-                loop {
-                    interval.tick().await;
-                    if cancel.is_cancelled() {
-                        break;
-                    }
-                    let processed =
-                        (progress.load(Ordering::Relaxed) as f64 / TARGET_SAMPLE_RATE).min(total);
-                    let percent = if total > 0.0 {
-                        ((processed / total) * 100.0) as u32
-                    } else {
-                        0
-                    };
-                    let _ = store
-                        .update(&id, Box::new(move |j| j.processed_seconds = processed))
-                        .await;
-                    broadcast_event(
-                        &*store,
-                        &id,
-                        JobEvent::Progress {
-                            percent,
-                            processed_seconds: processed,
-                        },
-                    )
-                    .await;
-                }
-            })
-        };
-
-        // Write-once sink for *why* speakers were or were not labeled. Armed
-        // only when diarization was requested, matching the synchronous path:
-        // an unarmed sink makes the engine record nothing.
-        let diar_sink: Option<
-            Arc<std::sync::OnceLock<gigastt_core::inference::DiarizationOutcome>>,
-        > = (params.diarization == Some(true)).then(|| Arc::new(std::sync::OnceLock::new()));
-
-        // Check out a triplet from the batch pool and run inference.
-        // This is wrapped in its own async block so the progress updater is
-        // always cancelled and awaited before the function returns, even on
-        // the early-return error paths below.
-        let inference_result: anyhow::Result<gigastt_core::inference::TranscribeResult> = async {
-            let guard = tokio::time::timeout(
-                std::time::Duration::from_secs(limits.pool_checkout_timeout_secs),
-                engine.pool_for_batch().checkout(),
-            )
-            .await
-            .map_err(|_| anyhow::anyhow!("pool checkout timed out"))?
-            .map_err(|_| anyhow::anyhow!("pool closed"))?;
-            let mut reservation = guard.into_owned();
-
-            let inference_timeout_secs = limits.inference_timeout_secs;
-            let engine_for_inference = engine.clone();
-            let body_for_inference = body.clone();
-            let span = tracing::Span::current();
-            let file_opts = super::file_transcribe::FileTranscribeOpts {
-                overrides,
-                hotwords,
-                split_channels: params.channels.as_deref() == Some("split"),
-                diarization: params.diarization == Some(true),
-                raw_codec: None,
-                abort: Some(abort.clone()),
-                progress: Some(progress.clone()),
-                // Same sink the synchronous endpoint uses, so an async job that
-                // produces no speaker labels can say why instead of returning a
-                // transcript with silently empty speaker fields. Only armed when
-                // diarization was actually requested; otherwise the engine
-                // records nothing and the response shape is untouched.
-                diarization_outcome: diar_sink.clone(),
-                max_audio_secs: limits.max_audio_secs_opt(),
-            };
-            let handle = tokio::task::spawn_blocking(move || {
-                let _enter = span.enter();
-                let r = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-                    super::file_transcribe::run_file_transcribe_blocking(
-                        &engine_for_inference,
-                        body_for_inference,
-                        &mut reservation,
-                        &file_opts,
-                    )
-                }));
-                match r {
-                    Ok(result) => result,
-                    Err(_) => Err(gigastt_core::error::GigasttError::Inference {
-                        source: anyhow::anyhow!("Inference thread panicked").into(),
-                    }),
-                }
-            });
-
-            // The inference timeout is a no-progress watchdog (shared with the
-            // REST path): the deadline resets on each completed window, so a long
-            // file making steady progress never trips, while a genuinely stalled
-            // run still fails with the deterministic `inference_timeout` marker.
-            // A fired shutdown also flips `abort`, freeing the triplet mid-drain.
-            match super::file_transcribe::await_transcription_watchdog(
-                handle,
-                &progress,
-                &abort,
-                inference_timeout_secs,
-                &shutdown,
-            )
-            .await
-            {
-                super::file_transcribe::WatchdogOutcome::Joined(r) => r
-                    .map_err(|e| anyhow::anyhow!("spawn_blocking join error: {e}"))?
-                    .map_err(anyhow::Error::from),
-                super::file_transcribe::WatchdogOutcome::TimedOut => {
-                    Err(anyhow::anyhow!("inference_timeout"))
-                }
-            }
-        }
-        .await;
-
-        progress_cancel.cancel();
-        let _ = progress_handle.await;
-
-        // Record the outcome on the job before returning: the worker stores the
-        // transcript after this call, and `GET /v1/jobs/{id}/result` reads both
-        // together. Written even on a failed run — the sink is empty then, so
-        // this is a no-op rather than a stale value.
-        if let Some(outcome) = diar_sink.as_ref().and_then(|s| s.get().copied()) {
-            let _ = store
-                .update(id, Box::new(move |j| j.diarization = Some(outcome)))
-                .await;
-        }
-
-        inference_result
-    }
-}
-
-#[cfg(test)]
-impl InMemoryJobStore {
-    /// Shift a job's `updated_at` back in time for TTL eviction tests.
-    pub async fn backdate(&self, id: &str, seconds: f64) {
-        let mut jobs = self.jobs.lock();
-        if let Some(job) = jobs.get_mut(id) {
-            job.updated_at -= seconds;
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn test_limits() -> RuntimeLimits {
-        RuntimeLimits {
-            jobs_enabled: true,
-            jobs_ttl_secs: 3600,
-            jobs_max: 10,
-            jobs_retry: 0,
-            ..RuntimeLimits::default()
-        }
-    }
-
-    #[tokio::test]
-    async fn test_in_memory_store_crud() {
-        let store = InMemoryJobStore::new(test_limits());
-        let id = store
-            .create(Job::queued(
-                Bytes::from_static(b"a"),
-                ExportParams::default(),
-            ))
-            .await
-            .unwrap();
-        let job = store.get(&id).await.unwrap().unwrap();
-        assert!(matches!(job.status, JobStatus::Queued));
-        store
-            .update(&id, Box::new(|j| j.status = JobStatus::Processing))
-            .await
-            .unwrap();
-        let job = store.get(&id).await.unwrap().unwrap();
-        assert!(matches!(job.status, JobStatus::Processing));
-    }
-
-    #[tokio::test]
-    async fn test_store_fifo_order() {
-        let store = InMemoryJobStore::new(test_limits());
-        let id1 = store
-            .create(Job::queued(
-                Bytes::from_static(b"1"),
-                ExportParams::default(),
-            ))
-            .await
-            .unwrap();
-        let id2 = store
-            .create(Job::queued(
-                Bytes::from_static(b"2"),
-                ExportParams::default(),
-            ))
-            .await
-            .unwrap();
-        assert_eq!(store.next_queued().await.unwrap(), Some(id1.clone()));
-        // id1 is still queued in the store; next_queued returned it but did not
-        // change its status. Simulate another worker trying to pop while id1
-        // is still queued: it should see id1 again because status is still Queued.
-        // Mark id1 processing and then next should be id2.
-        store
-            .update(&id1, Box::new(|j| j.status = JobStatus::Processing))
-            .await
-            .unwrap();
-        assert_eq!(store.next_queued().await.unwrap(), Some(id2));
-    }
-
-    #[tokio::test]
-    async fn test_store_capacity_limit() {
-        let limits = RuntimeLimits {
-            jobs_max: 1,
-            ..test_limits()
-        };
-        let store = InMemoryJobStore::new(limits);
-        store
-            .create(Job::queued(
-                Bytes::from_static(b"a"),
-                ExportParams::default(),
-            ))
-            .await
-            .unwrap();
-        assert!(store.is_full().await);
-        let result = store
-            .create(Job::queued(
-                Bytes::from_static(b"b"),
-                ExportParams::default(),
-            ))
-            .await;
-        assert!(result.is_err());
-    }
-
-    #[tokio::test]
-    async fn test_store_byte_budget_backpressures_under_count_limit() {
-        // Count limit is generous (10) but the byte budget is tiny: a store
-        // holding one upload already at/over the byte budget must report full
-        // and reject the next submission, even though it holds 1 << 10 jobs.
-        let limits = RuntimeLimits {
-            jobs_max: 10,
-            jobs_max_bytes: 8,
-            ..test_limits()
-        };
-        let store = InMemoryJobStore::new(limits);
-        // First upload (11 bytes) is admitted: like the count cap admitting the
-        // Nth job, the budget is checked against the bytes already resident (0).
-        store
-            .create(Job::queued(
-                Bytes::from_static(b"audio-bytes"),
-                ExportParams::default(),
-            ))
-            .await
-            .unwrap();
-        // 11 resident bytes now exceed the 8-byte budget, so the store is full
-        // by bytes despite being far below jobs_max, and the next create fails.
-        assert!(store.is_full().await);
-        let result = store
-            .create(Job::queued(
-                Bytes::from_static(b"more"),
-                ExportParams::default(),
-            ))
-            .await;
-        assert!(result.is_err());
-    }
-
-    #[tokio::test]
-    async fn test_store_byte_budget_released_when_job_terminal() {
-        // A terminal job releases its body, so those bytes stop counting against
-        // the budget and the queue accepts new work again.
-        let limits = RuntimeLimits {
-            jobs_max: 10,
-            jobs_max_bytes: 8,
-            ..test_limits()
-        };
-        let store = InMemoryJobStore::new(limits);
-        let id = store
-            .create(Job::queued(
-                Bytes::from_static(b"audio-bytes"),
-                ExportParams::default(),
-            ))
-            .await
-            .unwrap();
-        assert!(store.is_full().await);
-        // Release the body the way the worker does at a terminal state.
-        store
-            .update(
-                &id,
-                Box::new(|j| {
-                    j.status = JobStatus::Done;
-                    j.body = Bytes::new();
-                }),
-            )
-            .await
-            .unwrap();
-        assert!(!store.is_full().await);
-    }
-
-    #[tokio::test]
-    async fn test_store_is_full_evicts_expired() {
-        let limits = RuntimeLimits {
-            jobs_ttl_secs: 1,
-            jobs_max: 1,
-            ..test_limits()
-        };
-        let store = InMemoryJobStore::new(limits);
-        let id = store
-            .create(Job::queued(
-                Bytes::from_static(b"a"),
-                ExportParams::default(),
-            ))
-            .await
-            .unwrap();
-        store
-            .update(&id, Box::new(|j| j.status = JobStatus::Done))
-            .await
-            .unwrap();
-        store.backdate(&id, 2.0).await;
-        // is_full should evict the expired terminal job and report capacity.
-        assert!(!store.is_full().await);
-    }
-
-    #[tokio::test]
-    async fn test_store_ttl_eviction() {
-        let limits = RuntimeLimits {
-            jobs_ttl_secs: 1,
-            jobs_max: 10,
-            ..test_limits()
-        };
-        let store = InMemoryJobStore::new(limits);
-        let id = store
-            .create(Job::queued(
-                Bytes::from_static(b"a"),
-                ExportParams::default(),
-            ))
-            .await
-            .unwrap();
-        store
-            .update(&id, Box::new(|j| j.status = JobStatus::Done))
-            .await
-            .unwrap();
-        // Backdate the job by more than the 1-second TTL.
-        store.backdate(&id, 2.0).await;
-        // Creating a new job should evict the expired one.
-        store
-            .create(Job::queued(
-                Bytes::from_static(b"b"),
-                ExportParams::default(),
-            ))
-            .await
-            .unwrap();
-        assert!(store.get(&id).await.unwrap().is_none());
-    }
-
-    #[derive(Clone)]
-    struct MockExecutor {
-        results: Arc<Mutex<Vec<anyhow::Result<gigastt_core::inference::TranscribeResult>>>>,
-        delay_ms: u64,
-    }
-
-    impl JobExecution for MockExecutor {
-        async fn execute(
-            &self,
-            _id: &str,
-            _store: Arc<dyn JobStore>,
-            body: Bytes,
-            _params: ExportParams,
-        ) -> anyhow::Result<gigastt_core::inference::TranscribeResult> {
-            if self.delay_ms > 0 {
-                tokio::time::sleep(std::time::Duration::from_millis(self.delay_ms)).await;
-            }
-            let mut results = self.results.lock();
-            let result = results.remove(0);
-            // Use body length to make failures deterministic per test.
-            let _ = body.len();
-            result
-        }
-    }
-
-    fn ok_result() -> anyhow::Result<gigastt_core::inference::TranscribeResult> {
-        Ok(gigastt_core::inference::TranscribeResult {
-            text: "ok".into(),
-            words: vec![],
-            duration_s: 1.0,
-            confidence: None,
-        })
-    }
-
-    #[tokio::test]
-    async fn test_queue_runs_jobs_in_fifo_order() {
-        let limits = test_limits();
-        let store: Arc<dyn JobStore> = Arc::new(InMemoryJobStore::new(limits));
-        let executor = MockExecutor {
-            results: Arc::new(Mutex::new(vec![ok_result(), ok_result()])),
-            delay_ms: 0,
-        };
-        let queue = JobQueue::new(
-            store.clone(),
-            2,
-            0,
-            tokio_util::sync::CancellationToken::new(),
-        );
-        let tracker = tokio_util::task::TaskTracker::new();
-        queue.spawn(executor, &tracker);
-
-        let id1 = store
-            .create(Job::queued(
-                Bytes::from_static(b"1"),
-                ExportParams::default(),
-            ))
-            .await
-            .unwrap();
-        let id2 = store
-            .create(Job::queued(
-                Bytes::from_static(b"2"),
-                ExportParams::default(),
-            ))
-            .await
-            .unwrap();
-
-        // Wait for both to finish.
-        for _ in 0..50 {
-            let j1 = store.get(&id1).await.unwrap().unwrap();
-            let j2 = store.get(&id2).await.unwrap().unwrap();
-            if matches!(j1.status, JobStatus::Done) && matches!(j2.status, JobStatus::Done) {
-                break;
-            }
-            tokio::time::sleep(std::time::Duration::from_millis(50)).await;
-        }
-
-        assert!(matches!(
-            store.get(&id1).await.unwrap().unwrap().status,
-            JobStatus::Done
-        ));
-        assert!(matches!(
-            store.get(&id2).await.unwrap().unwrap().status,
-            JobStatus::Done
-        ));
-    }
-
-    #[tokio::test]
-    async fn test_queue_retry_then_fail() {
-        let limits = RuntimeLimits {
-            jobs_retry: 2,
-            ..test_limits()
-        };
-        let store: Arc<dyn JobStore> = Arc::new(InMemoryJobStore::new(limits));
-        let executor = MockExecutor {
-            // Panic is the retryable failure class (inference_timeout is not),
-            // so the retry path is exercised with a panic marker.
-            results: Arc::new(Mutex::new(vec![
-                Err(anyhow::anyhow!("worker thread panicked")),
-                Err(anyhow::anyhow!("worker thread panicked")),
-                Err(anyhow::anyhow!("worker thread panicked")),
-            ])),
-            delay_ms: 0,
-        };
-        let queue = JobQueue::new(
-            store.clone(),
-            1,
-            2,
-            tokio_util::sync::CancellationToken::new(),
-        );
-        let tracker = tokio_util::task::TaskTracker::new();
-        queue.spawn(executor, &tracker);
-
-        let id = store
-            .create(Job::queued(
-                Bytes::from_static(b"x"),
-                ExportParams::default(),
-            ))
-            .await
-            .unwrap();
-
-        for _ in 0..50 {
-            let job = store.get(&id).await.unwrap().unwrap();
-            if matches!(job.status, JobStatus::Failed) {
-                break;
-            }
-            tokio::time::sleep(std::time::Duration::from_millis(50)).await;
-        }
-
-        let job = store.get(&id).await.unwrap().unwrap();
-        assert!(matches!(job.status, JobStatus::Failed));
-        assert_eq!(job.attempts, 3);
-    }
-
-    #[tokio::test]
-    async fn test_queue_cancellation_discards_result() {
-        let store: Arc<dyn JobStore> = Arc::new(InMemoryJobStore::new(test_limits()));
-        let executor = MockExecutor {
-            results: Arc::new(Mutex::new(vec![ok_result()])),
-            delay_ms: 300,
-        };
-        let queue = JobQueue::new(
-            store.clone(),
-            1,
-            0,
-            tokio_util::sync::CancellationToken::new(),
-        );
-        let tracker = tokio_util::task::TaskTracker::new();
-        queue.spawn(executor, &tracker);
-
-        let id = store
-            .create(Job::queued(
-                Bytes::from_static(b"x"),
-                ExportParams::default(),
-            ))
-            .await
-            .unwrap();
-
-        // Wait until processing, then cancel while the executor is still running.
-        for _ in 0..50 {
-            let job = store.get(&id).await.unwrap().unwrap();
-            if matches!(job.status, JobStatus::Processing) {
-                break;
-            }
-            tokio::time::sleep(std::time::Duration::from_millis(20)).await;
-        }
-        store
-            .update(&id, Box::new(|j| j.status = JobStatus::Cancelled))
-            .await
-            .unwrap();
-
-        tokio::time::sleep(std::time::Duration::from_millis(500)).await;
-        let job = store.get(&id).await.unwrap().unwrap();
-        assert!(matches!(job.status, JobStatus::Cancelled));
-        assert!(job.result.is_none());
-    }
-
-    #[tokio::test]
-    async fn test_queue_cancel_all_queued_on_shutdown() {
-        let store: Arc<dyn JobStore> = Arc::new(InMemoryJobStore::new(test_limits()));
-        let token = tokio_util::sync::CancellationToken::new();
-        let queue = JobQueue::new(store.clone(), 1, 0, token.clone());
-        // Don't spawn workers so jobs stay queued.
-
-        let id = store
-            .create(Job::queued(
-                Bytes::from_static(b"x"),
-                ExportParams::default(),
-            ))
-            .await
-            .unwrap();
-
-        token.cancel();
-        queue.cancel_all_queued().await;
-
-        let job = store.get(&id).await.unwrap().unwrap();
-        assert!(matches!(job.status, JobStatus::Cancelled));
-    }
-
-    #[test]
-    fn test_sanitize_job_error_maps_known_errors() {
-        assert_eq!(
-            sanitize_job_error(&anyhow::anyhow!("inference_timeout")),
-            "Inference timed out."
-        );
-        assert_eq!(
-            sanitize_job_error(&anyhow::anyhow!("Invalid audio: unsupported format")),
-            "Failed to decode audio file. Check format."
-        );
-        assert_eq!(
-            sanitize_job_error(&anyhow::anyhow!("some internal onnx path /foo/bar")),
-            "Transcription failed."
-        );
-        // A typed AudioTooLong (arrives via `anyhow::Error::from`) surfaces the
-        // observed/limit seconds so a client can tell "too long" from "corrupt".
-        let too_long: anyhow::Error = gigastt_core::error::GigasttError::AudioTooLong {
-            observed_secs: 4000.0,
-            limit_secs: 1800.0,
-        }
-        .into();
-        assert_eq!(
-            sanitize_job_error(&too_long),
-            "Audio too long: 4000s exceeds the maximum of 1800s."
-        );
-    }
-
-    #[test]
-    fn test_is_retryable_error_recognizes_transient_failures() {
-        // A panic is transient — a fresh worker may succeed. An inference_timeout
-        // is deterministic, so it is NOT retryable. A decode error never is.
-        assert!(is_retryable_error(&anyhow::anyhow!(
-            "worker thread panicked"
-        )));
-        assert!(!is_retryable_error(&anyhow::anyhow!("inference_timeout")));
-        assert!(!is_retryable_error(&anyhow::anyhow!("Invalid audio")));
-    }
-
-    #[tokio::test]
-    async fn test_store_get_missing_returns_none() {
-        let store = InMemoryJobStore::new(test_limits());
-        assert!(store.get("no-such-id").await.unwrap().is_none());
-    }
-
-    #[tokio::test]
-    async fn test_store_update_missing_returns_error() {
-        let store = InMemoryJobStore::new(test_limits());
-        assert!(
-            store
-                .update("no-such-id", Box::new(|j| j.status = JobStatus::Done))
-                .await
-                .is_err()
-        );
-    }
-
-    #[tokio::test]
-    async fn test_broadcast_event_prunes_dead_channels() {
-        let store: Arc<dyn JobStore> = Arc::new(InMemoryJobStore::new(test_limits()));
-        let id = store
-            .create(Job::queued(
-                Bytes::from_static(b"x"),
-                ExportParams::default(),
-            ))
-            .await
-            .unwrap();
-
-        // Add a live channel and a channel that will be dropped before broadcast.
-        let (live_tx, mut live_rx) = tokio::sync::mpsc::unbounded_channel::<JobEvent>();
-        {
-            let (dead_tx, _dead_rx) = tokio::sync::mpsc::unbounded_channel::<JobEvent>();
-            store
-                .update(
-                    &id,
-                    Box::new(move |j| {
-                        j.event_channels.push(live_tx);
-                        j.event_channels.push(dead_tx);
-                    }),
-                )
-                .await
-                .unwrap();
-        }
-
-        broadcast_event(
-            &*store,
-            &id,
-            JobEvent::Progress {
-                percent: 50,
-                processed_seconds: 1.0,
-            },
-        )
-        .await;
-
-        let job = store.get(&id).await.unwrap().unwrap();
-        assert_eq!(job.event_channels.len(), 1);
-        assert!(matches!(live_rx.try_recv(), Ok(JobEvent::Progress { .. })));
-    }
-
-    #[tokio::test]
-    async fn test_queue_concurrency_clamped_to_one() {
-        let store: Arc<dyn JobStore> = Arc::new(InMemoryJobStore::new(test_limits()));
-        let executor = MockExecutor {
-            results: Arc::new(Mutex::new(vec![ok_result(), ok_result()])),
-            delay_ms: 0,
-        };
-        // Pass 0; the queue must still spawn at least one worker.
-        let queue = JobQueue::new(
-            store.clone(),
-            0,
-            0,
-            tokio_util::sync::CancellationToken::new(),
-        );
-        let tracker = tokio_util::task::TaskTracker::new();
-        queue.spawn(executor, &tracker);
-
-        let id = store
-            .create(Job::queued(
-                Bytes::from_static(b"x"),
-                ExportParams::default(),
-            ))
-            .await
-            .unwrap();
-
-        for _ in 0..50 {
-            let job = store.get(&id).await.unwrap().unwrap();
-            if matches!(job.status, JobStatus::Done) {
-                break;
-            }
-            tokio::time::sleep(std::time::Duration::from_millis(50)).await;
-        }
-
-        assert!(matches!(
-            store.get(&id).await.unwrap().unwrap().status,
-            JobStatus::Done
-        ));
-    }
-
-    #[tokio::test]
-    async fn test_queue_non_retryable_error_fails_immediately() {
-        let store: Arc<dyn JobStore> = Arc::new(InMemoryJobStore::new(test_limits()));
-        let executor = MockExecutor {
-            results: Arc::new(Mutex::new(vec![Err(anyhow::anyhow!(
-                "Invalid audio: bad header"
-            ))])),
-            delay_ms: 0,
-        };
-        let queue = JobQueue::new(
-            store.clone(),
-            1,
-            3,
-            tokio_util::sync::CancellationToken::new(),
-        );
-        let tracker = tokio_util::task::TaskTracker::new();
-        queue.spawn(executor, &tracker);
-
-        let id = store
-            .create(Job::queued(
-                Bytes::from_static(b"x"),
-                ExportParams::default(),
-            ))
-            .await
-            .unwrap();
-
-        for _ in 0..50 {
-            let job = store.get(&id).await.unwrap().unwrap();
-            if matches!(job.status, JobStatus::Failed) {
-                break;
-            }
-            tokio::time::sleep(std::time::Duration::from_millis(50)).await;
-        }
-
-        let job = store.get(&id).await.unwrap().unwrap();
-        assert!(matches!(job.status, JobStatus::Failed));
-        assert_eq!(job.attempts, 1);
-    }
-
-    #[tokio::test]
-    async fn test_queue_timeout_fails_once() {
-        // A deterministic inference_timeout must fail on its first and only
-        // attempt — even with retry budget to spare — so a too-slow file can't
-        // burn (max_retries + 1) × the timeout while the queue waits.
-        let store: Arc<dyn JobStore> = Arc::new(InMemoryJobStore::new(test_limits()));
-        let executor = MockExecutor {
-            // Second result is queued but must never be consumed: no retry.
-            results: Arc::new(Mutex::new(vec![
-                Err(anyhow::anyhow!("inference_timeout")),
-                ok_result(),
-            ])),
-            delay_ms: 0,
-        };
-        // Generous retry budget (3) — the point is that timeout ignores it.
-        let queue = JobQueue::new(
-            store.clone(),
-            1,
-            3,
-            tokio_util::sync::CancellationToken::new(),
-        );
-        let tracker = tokio_util::task::TaskTracker::new();
-        queue.spawn(executor, &tracker);
-
-        let id = store
-            .create(Job::queued(
-                Bytes::from_static(b"x"),
-                ExportParams::default(),
-            ))
-            .await
-            .unwrap();
-
-        for _ in 0..50 {
-            let job = store.get(&id).await.unwrap().unwrap();
-            if matches!(job.status, JobStatus::Failed) {
-                break;
-            }
-            tokio::time::sleep(std::time::Duration::from_millis(50)).await;
-        }
-
-        let job = store.get(&id).await.unwrap().unwrap();
-        assert!(matches!(job.status, JobStatus::Failed));
-        assert_eq!(
-            job.attempts, 1,
-            "inference_timeout must fail once, not retry"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_queue_retry_boundary_exhausted_after_max_retries() {
-        // max_retries=1 means one retry is allowed: attempts 1 (retry), 2 (fail).
-        let limits = RuntimeLimits {
-            jobs_retry: 1,
-            ..test_limits()
-        };
-        let store: Arc<dyn JobStore> = Arc::new(InMemoryJobStore::new(limits));
-        let executor = MockExecutor {
-            // Retry path is exercised with a panic marker (the retryable class);
-            // inference_timeout is deterministic and fails once.
-            results: Arc::new(Mutex::new(vec![
-                Err(anyhow::anyhow!("worker thread panicked")),
-                Err(anyhow::anyhow!("worker thread panicked")),
-            ])),
-            delay_ms: 0,
-        };
-        let queue = JobQueue::new(
-            store.clone(),
-            1,
-            1,
-            tokio_util::sync::CancellationToken::new(),
-        );
-        let tracker = tokio_util::task::TaskTracker::new();
-        queue.spawn(executor, &tracker);
-
-        let id = store
-            .create(Job::queued(
-                Bytes::from_static(b"x"),
-                ExportParams::default(),
-            ))
-            .await
-            .unwrap();
-
-        for _ in 0..50 {
-            let job = store.get(&id).await.unwrap().unwrap();
-            if matches!(job.status, JobStatus::Failed) {
-                break;
-            }
-            tokio::time::sleep(std::time::Duration::from_millis(50)).await;
-        }
-
-        let job = store.get(&id).await.unwrap().unwrap();
-        assert!(matches!(job.status, JobStatus::Failed));
-        assert_eq!(job.attempts, 2);
-    }
-
-    #[test]
-    fn test_job_status_response_percent() {
-        let mut job = Job::queued(Bytes::new(), ExportParams::default());
-        job.id = "test".into();
-        job.total_seconds = 10.0;
-        job.processed_seconds = 3.5;
-        job.status = JobStatus::Processing;
-        let resp = job_status_response(&job);
-        assert_eq!(resp.percent, 35);
-        assert_eq!(resp.processed_seconds, 3.5);
-    }
-
-    #[tokio::test]
-    async fn test_queue_done_job_releases_body() {
-        let store: Arc<dyn JobStore> = Arc::new(InMemoryJobStore::new(test_limits()));
-        let executor = MockExecutor {
-            results: Arc::new(Mutex::new(vec![ok_result()])),
-            delay_ms: 0,
-        };
-        let queue = JobQueue::new(
-            store.clone(),
-            1,
-            0,
-            tokio_util::sync::CancellationToken::new(),
-        );
-        let tracker = tokio_util::task::TaskTracker::new();
-        queue.spawn(executor, &tracker);
-
-        let id = store
-            .create(Job::queued(
-                Bytes::from_static(b"audio-bytes"),
-                ExportParams::default(),
-            ))
-            .await
-            .unwrap();
-
-        for _ in 0..50 {
-            let job = store.get(&id).await.unwrap().unwrap();
-            if matches!(job.status, JobStatus::Done) {
-                break;
-            }
-            tokio::time::sleep(std::time::Duration::from_millis(50)).await;
-        }
-
-        let job = store.get(&id).await.unwrap().unwrap();
-        assert!(matches!(job.status, JobStatus::Done));
-        assert!(job.body.is_empty());
-    }
-
-    /// Records the body length seen by each attempt so tests can verify the
-    /// upload survives retries and is only released at a terminal state.
-    #[derive(Clone)]
-    struct BodyLenRecorder {
-        lens: Arc<Mutex<Vec<usize>>>,
-    }
-
-    impl JobExecution for BodyLenRecorder {
-        async fn execute(
-            &self,
-            _id: &str,
-            _store: Arc<dyn JobStore>,
-            body: Bytes,
-            _params: ExportParams,
-        ) -> anyhow::Result<gigastt_core::inference::TranscribeResult> {
-            self.lens.lock().push(body.len());
-            // Retryable failure class, so the body must survive to the next
-            // attempt and only be released at the terminal state.
-            Err(anyhow::anyhow!("worker thread panicked"))
-        }
-    }
-
-    #[tokio::test]
-    async fn test_queue_failed_job_releases_body_after_retries() {
-        // max_retries=1 → two attempts, then terminal Failed.
-        let limits = RuntimeLimits {
-            jobs_retry: 1,
-            ..test_limits()
-        };
-        let store: Arc<dyn JobStore> = Arc::new(InMemoryJobStore::new(limits));
-        let lens = Arc::new(Mutex::new(Vec::new()));
-        let executor = BodyLenRecorder { lens: lens.clone() };
-        let queue = JobQueue::new(
-            store.clone(),
-            1,
-            1,
-            tokio_util::sync::CancellationToken::new(),
-        );
-        let tracker = tokio_util::task::TaskTracker::new();
-        queue.spawn(executor, &tracker);
-
-        let id = store
-            .create(Job::queued(
-                Bytes::from_static(b"audio-bytes"),
-                ExportParams::default(),
-            ))
-            .await
-            .unwrap();
-
-        for _ in 0..50 {
-            let job = store.get(&id).await.unwrap().unwrap();
-            if matches!(job.status, JobStatus::Failed) {
-                break;
-            }
-            tokio::time::sleep(std::time::Duration::from_millis(50)).await;
-        }
-
-        let job = store.get(&id).await.unwrap().unwrap();
-        assert!(matches!(job.status, JobStatus::Failed));
-        // Both attempts received the full upload; the body is released only
-        // once the job reaches a terminal state.
-        assert_eq!(*lens.lock(), vec![11, 11]);
-        assert!(job.body.is_empty());
-    }
-
-    #[test]
-    fn test_subscribe_prunes_closed_channels() {
-        let mut job = Job::queued(Bytes::new(), ExportParams::default());
-        let (dead_tx, dead_rx) = tokio::sync::mpsc::unbounded_channel::<JobEvent>();
-        job.event_channels.push(dead_tx);
-        drop(dead_rx);
-
-        let (live_tx, _live_rx) = tokio::sync::mpsc::unbounded_channel::<JobEvent>();
-        job.subscribe(live_tx);
-
-        assert_eq!(job.event_channels.len(), 1);
-    }
-
-    #[test]
-    fn test_subscribe_evicts_oldest_at_cap() {
-        let mut job = Job::queued(Bytes::new(), ExportParams::default());
-        let mut rxs = Vec::new();
-        for _ in 0..MAX_JOB_EVENT_SUBSCRIBERS {
-            let (tx, rx) = tokio::sync::mpsc::unbounded_channel::<JobEvent>();
-            job.subscribe(tx);
-            rxs.push(rx);
-        }
-        assert_eq!(job.event_channels.len(), MAX_JOB_EVENT_SUBSCRIBERS);
-
-        let (tx, rx) = tokio::sync::mpsc::unbounded_channel::<JobEvent>();
-        job.subscribe(tx);
-        rxs.push(rx);
-
-        // The list stays capped; the oldest listener was evicted, so its
-        // stream is disconnected.
-        assert_eq!(job.event_channels.len(), MAX_JOB_EVENT_SUBSCRIBERS);
-        assert!(matches!(
-            rxs[0].try_recv(),
-            Err(tokio::sync::mpsc::error::TryRecvError::Disconnected)
-        ));
-    }
-}
